Adds a new @version docblock to all newly generated blocks
By adding this to all newly generated blocks we now have the options of knowing which vwersion was used to generate each block. This could help in future when breaking changes are introduced as we have a universal way of targeting blocks and knowing which versions they were created using. This will allow us to build more specific rules.